requirements.txt
# Function dependencies, for example:
# package>=version
functions-framework==1.2.0
inject==3.4.0
PyYAML==3.13